+++ crates/core/src/resources/world/mod.rs
